requirements.txt
paho-mqtt
prometheus-client
pyyaml
confuse

